pdf_drive_crawler/requirements.txt
pandas
numpy
PdfReader
tabula-py